Move to_id_str into poke_env.data
