+++ src/poke_env/data.py
@@ -37,6 +37,16 @@ def _compute_type_chart(chart_path: str) -> Dict[str, Dict[str, float]]:
     return type_chart
 
 
+def to_id_str(name: str) -> str:
+    """Converts a full-name to its corresponding id string.
+    :param name: The name to convert.
+    :type name: str
+    :return: The corresponding id string.
+    :rtype: str
+    """
+    return "".join(char for char in name if char.isalnum()).lower()
+
+
 _TYPE_CHART_PATH: str = os.path.join(
     os.path.dirname(os.path.realpath(__file__)), "data", "typeChart.json"
 )
